+++ p2/yylex.c
@@ -59,7 +59,7 @@ lexgetc()
 		c = peekc;
 		peekc = 0;
 		return(c);
-	}else if (((int) c = getc(lexin))==EOF){
+	}else if (( c = getc(lexin))==EOF){
 	      if(in_index!=0){
 		      in_index--;
 		      lexin = in_files[in_index].in_fdes;
